fixes #63
- Added test in echo for reading env variables
- finished reorganizing mock classes into separate files
