+++ packages/assemblyscript/assembly/__tests__/wasa.include.ts
@@ -2,26 +2,7 @@
 import { Console, fs, Process, CommandLine, fd, FileDescriptor } from '../wasa/mock/index';
 import { Wasi } from "../wasi";
 
-beforeAll(() => {
-
-    fs.fs.init();
-
-})
 /** This file is included with tests so that the default globals are set up properly. */
-//@ts-ignore
-// @global
-// const Console = _Console;
-// Console.stdin
-// Console.stdout
-// Console.stderr
-
-// let x: fd = 0;
-// //@ts-ignore
-// const Process = _Process;
-//@ts-ignore
-// fs.openDirectory("/", fs.openForRead("/.file"))
-// CommandLine.reset()
-// //@ts-ignore
-// const fd = usize;
-// //@ts-ignore
-// const CommandLine = _CommandLine;
+beforeAll(() => {
+    fs.fs.init();
+})
