+++ packages/assemblyscript/assembly/wasa/mock/performance.ts
@@ -1,0 +1,9 @@
+
+export class Performance {
+  static now(): f64 {
+    let time_ptr = memory.allocate(8);
+    let res_ts = load<u64>(time_ptr);
+    memory.free(time_ptr);
+    return res_ts as f64;
+  }
+}
