+++ packages/assemblyscript/assembly/wasi/index.ts
@@ -1,4 +1,5 @@
-
+//@ts-ignore
+@global
 export namespace Wasi {
 
     // === Types ======================================================================================
